Update readline inputrc
From
the (manual)[https://www.gnu.org/software/bash/manual/html_node/Readline-Init-File-Syntax.html]

blink-matching-paren
If set to ‘on’, Readline attempts to briefly move the cursor to an
opening parenthesis when a closing parenthesis is inserted. The
default is ‘off’.

colored-completion-prefix
If set to ‘on’, when listing completions, Readline displays the common
prefix of the set of possible completions using a different color. The
color definitions are taken from the value of the LS_COLORS
environment variable. If there is a color definition in LS_COLORS for
the custom suffix ‘readline-colored-completion-prefix’, Readline uses
this color for the common prefix instead of its default. The default
is ‘off’.

enable-bracketed-paste
When set to ‘On’, Readline configures the terminal to insert each
paste into the editing buffer as a single string of characters,
instead of treating each character as if it had been read from the
keyboard. This is called putting the terminal into bracketed paste
mode; it prevents Readline from executing any editing commands bound
to key sequences appearing in the pasted text. The default is ‘On’.